refactor: rename webview-ui to webview
This can contain more than just UI content (and it should: e.g., IPC
definitions, etc...).
