+++ src/webviews/README.md
@@ -1,4 +1,4 @@
-# `webview-ui` Directory
+# `webview` Directory
 
 This directory contains all of the code that will be executed within the webview context. It can be thought of as the place where all the "frontend" code of a webview is contained.
