+++ src/webviews/components/SearchResults/EntryHeader.tsx
