+++ src/webviews/components/SearchResults/PathHeader.tsx
