+++ src/webviews/components/SearchResults/SearchResultEntry.tsx
